+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## **v1.1.0** - 2025-04-08
 ### ➕ Added
-* [web-audio] `autoplaySpots` options
+* [web-audio] `autoplaySounds` options
 * [web-audio] `ready` event
 
 ## **v1.0.3** - 2025-01-07